refactor: update imports, move styles to modules
