+++ components/organisms/Footer/Footer.module.scss
@@ -1,0 +1,5 @@
+.footer {
+  padding: 1rem;
+  text-align: center;
+  margin-top: auto;
+}
