+++ components/organisms/Footer/Footer.tsx
@@ -1,23 +1,17 @@
 import {FontAwesomeIcon} from '@fortawesome/react-fontawesome'
 import {faGithubAlt} from '@fortawesome/free-brands-svg-icons'
+import styles from './Footer.module.scss'
 
 const Footer = () => {
   return (
     <>
-      <footer className="footer">
+      <footer className={styles.footer}>
         <div className="container">
           @ 2022 - <a href="https://github.com/yrogovich/next-blog" target="_blank" rel="noreferrer">
           <FontAwesomeIcon icon={faGithubAlt}/> yrogovich
         </a>
         </div>
       </footer>
-      <style jsx>{`
-        .footer {
-          padding: 1rem;
-          text-align: center;
-          margin-top: auto;
-        }
-      `}</style>
     </>
   )
 }
