+++ components/organisms/Navbar/Navbar.module.scss
@@ -1,0 +1,15 @@
+.navbar {
+  padding: 2rem 0;
+}
+
+.row {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  gap: 1rem;
+}
+
+.logo {
+  font-size: 2.6rem;
+  font-weight: 700;
+}
